applications group

+++ terraform/01-kubernetes/dependencies.sh
@@ -8,6 +8,7 @@ sudo apt-get -y upgrade
 
 # Dist-upgrade all installed packages to their latest versions
 sudo apt-get -y dist-upgrade
+# TODO: i think this is bad practice because we don't reboot into the new kernel after
 
 # Install required packages
 sudo apt-get -y install acl htop psmisc netcat-traditional
@@ -16,6 +17,10 @@ sudo apt-get -y install acl htop psmisc netcat-traditional
 sudo apt-get -y autoremove
 
 # Install k9s
-wget https://github.com/derailed/k9s/releases/download/v0.40.5/k9s_linux_amd64.deb
-sudo apt install -y ./k9s_linux_*.deb
-rm k9s_linux_*.deb
+if ! command -v k9s &> /dev/null; then
+    wget https://github.com/derailed/k9s/releases/download/v0.40.5/k9s_linux_amd64.deb
+    sudo apt install -y ./k9s_linux_*.deb
+    rm k9s_linux_*.deb
+else
+    echo "k9s is already installed"
+fi
